requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
docker==5.0.3
docopt==0.6.2
Flask==2.0.2
halo==0.0.31
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
kafka-python==2.0.2
kazoo==2.8.0
keras==2.6.0
kiel==0.9.3
log-symbols==0.0.14
MarkupSafe==2.0.1
numpy==1.21.3
opencv-python==4.5.4.58
pi==0.1.2
Pillow==8.4.0
pipreqs==0.4.11
pybind11==2.8.0
pytz==2021.3
qrcode==7.3.1
requests==2.26.0
six==1.16.0
spinners==0.0.24
termcolor==1.1.0
tornado==6.1
urllib3==1.26.7
ushlex==0.99.1
websocket-client==1.2.1
websockets==10.0
Werkzeug==2.0.2
yarg==0.1.9
